feat: align proposer preferences api with latest spec (#9570)
Aligns our implementation with
https://github.com/ethereum/beacon-APIs/pull/608
- endpoint moved `/eth/v1/beacon/pool/proposer_preferences` -->
`/eth/v1/validator/proposer_preferences`
- drop the `GET /eth/v1/beacon/pool/proposer_preferences` endpoint, we
can reconsider adding a `/lodestar` endpoint for this if it would be
useful for debugging or propose it to the spec, but for now I don't
think it's that useful to have
- list capped at `(MIN_SEED_LOOKAHEAD + 1) * SLOTS_PER_EPOCH` per spec
- derive `Eth-Consensus-Version` header and event `version` from
proposal slot
