+++ packages/api/test/unit/beacon/testData/beacon.ts
@@ -119,10 +119,6 @@ export const testData: GenericServerTestCases<Endpoints> = {
     args: {slot: 1},
     res: {data: [ssz.gloas.PayloadAttestation.defaultValue()], meta: {version: ForkName.gloas}},
   },
-  getPoolProposerPreferences: {
-    args: {slot: 1},
-    res: {data: [ssz.gloas.SignedProposerPreferences.defaultValue()], meta: {version: ForkName.gloas}},
-  },
   getPoolAttesterSlashingsV2: {
     args: undefined,
     res: {data: [ssz.electra.AttesterSlashing.defaultValue()], meta: {version: ForkName.electra}},
@@ -165,10 +161,6 @@ export const testData: GenericServerTestCases<Endpoints> = {
   },
   submitPayloadAttestationMessages: {
     args: {payloadAttestationMessages: [ssz.gloas.PayloadAttestationMessage.defaultValue()]},
-    res: undefined,
-  },
-  submitSignedProposerPreferences: {
-    args: {signedProposerPreferences: [ssz.gloas.SignedProposerPreferences.defaultValue()]},
     res: undefined,
   },
