+++ packages/beacon-node/src/api/impl/beacon/pool/index.ts
@@ -1,7 +1,6 @@
 import {routes} from "@lodestar/api";
 import {ApplicationMethods} from "@lodestar/api/server";
 import {
-  ForkName,
   ForkPostElectra,
   ForkPreElectra,
   SYNC_COMMITTEE_SUBNET_SIZE,
@@ -17,15 +16,12 @@ import {
   GossipAction,
   PayloadAttestationError,
   PayloadAttestationErrorCode,
-  ProposerPreferencesError,
-  ProposerPreferencesErrorCode,
   SyncCommitteeError,
 } from "../../../../chain/errors/index.js";
 import {validateApiAttesterSlashing} from "../../../../chain/validation/attesterSlashing.js";
 import {validateApiBlsToExecutionChange} from "../../../../chain/validation/blsToExecutionChange.js";
 import {toElectraSingleAttestation, validateApiAttestation} from "../../../../chain/validation/index.js";
 import {validateApiPayloadAttestationMessage} from "../../../../chain/validation/payloadAttestationMessage.js";
-import {validateGossipProposerPreferences} from "../../../../chain/validation/proposerPreferences.js";
 import {validateApiProposerSlashing} from "../../../../chain/validation/proposerSlashing.js";
 import {validateApiSyncCommittee} from "../../../../chain/validation/syncCommittee.js";
 import {validateApiVoluntaryExit} from "../../../../chain/validation/voluntaryExit.js";
@@ -64,54 +60,6 @@ export function getBeaconPoolApi({
       }
 
       return {data: chain.payloadAttestationPool.getAll(slot), meta: {version: fork}};
-    },
-
-    async getPoolProposerPreferences({slot}) {
-      const fork = chain.config.getForkName(slot ?? chain.clock.currentSlot);
-      if (!isForkPostGloas(fork)) {
-        throw new ApiError(400, `Proposer preferences pool is not supported before Gloas fork=${fork}`);
-      }
-
-      return {data: chain.proposerPreferencesPool.getAll(slot), meta: {version: fork}};
-    },
-
-    async submitSignedProposerPreferences({signedProposerPreferences}) {
-      const failures: FailureList = [];
-
-      await Promise.all(
-        signedProposerPreferences.map(async (signed, i) => {
-          try {
-            await validateGossipProposerPreferences(chain, signed);
-
-            await network.publishProposerPreferences(signed);
-            chain.emitter.emit(routes.events.EventType.proposerPreferences, {
-              version: ForkName.gloas,
-              data: signed,
-            });
-          } catch (e) {
-            const logCtx = {
-              slot: signed.message.proposalSlot,
-              validatorIndex: signed.message.validatorIndex,
-              dependentRoot: toRootHex(signed.message.dependentRoot),
-            };
-
-            if (e instanceof ProposerPreferencesError && e.type.code === ProposerPreferencesErrorCode.ALREADY_KNOWN) {
-              logger.debug("Ignoring known signed proposer preferences", logCtx);
-              return;
-            }
-
-            failures.push({index: i, message: (e as Error).message});
-            logger.verbose(`Error on submitSignedProposerPreferences [${i}]`, logCtx, e as Error);
-            if (e instanceof ProposerPreferencesError && e.action === GossipAction.REJECT) {
-              chain.persistInvalidSszValue(ssz.gloas.SignedProposerPreferences, signed, "api_reject");
-            }
-          }
-        })
-      );
-
-      if (failures.length > 0) {
-        throw new IndexedError("Error processing signed proposer preferences", failures);
-      }
     },
 
     async getPoolAttesterSlashingsV2() {
