+++ packages/validator/src/services/proposerPreferences.ts
@@ -118,7 +118,7 @@ export class ProposerPreferencesService {
     }
 
     try {
-      await this.api.beacon.submitSignedProposerPreferences({signedProposerPreferences: batch});
+      await this.api.validator.submitProposerPreferences({signedProposerPreferences: batch});
       // Only mark as submitted after the API call succeeds; a thrown error leaves the
       // slot eligible for retry on the next tick.
       for (const {submission, slot: submittedSlot} of pending) {
